functions/orchestration-helpers/scheduling-function/requirements.txt
functions-framework==3.3.0
google-cloud-bigquery==3.11.4
google-cloud-logging
google-cloud-error-reporting
google-cloud-firestore
google-events